command to easily test effects, first part of effect xml loader

+++ src/fluorescence/ui/particles/particleemitterstate.hpp
@@ -39,7 +39,7 @@ class ParticleEmitterState {
 public:
     UnicodeString name_;
     
-    unsigned int emitFrequency_;
+    float emitFrequency_;
     
     boost::shared_ptr<StartLocationProvider> emittedStartLocationProvider_;
     boost::shared_ptr<MotionModel> emittedMotionModel_;
